Add RPC implementation using RabbitMQ: Order

+++ go/orders/models/models.go
@@ -10,6 +10,6 @@ type Orders struct {
 	EventID     uint           `json:"event_id"`
 	UserID      uint           `json:"user_id"`
 	Details     datatypes.JSON `json:"details" example:"{\"classA\": [1,2,3], \"classB\": [1]}"`
-	TotalAmount float32        `json:"total_amount"`
+	TotalAmount float64        `json:"total_amount"`
 	IsPaid      bool           `json:"is_paid"`
 }
